FastAPI Project with GS/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
gspread>=6.0.0
google-auth>=2.29.0
python-dotenv>=1.0.0
pydantic>=2.0.0
